debug

+++ build_images.sh
@@ -7,17 +7,19 @@ echo "NUMBER OF IMAGES TO BUILD: $folder_count"
 # get and loop through all folder names in fake_apps dir on the first hierarchical level
 find fake_apps/ -mindepth 1 -maxdepth 1 -type d | while read folder_name; do
 
-    # extract folder name from the file path
-    image_name=$(basename "$folder_name")
-
-    # remove existing Docker image from registry
-    docker rmi -f "$image_name"
+    # check if Docker image exists in registry
+    if docker images -q "$image_name" >/dev/null 2>&1; then
+        echo "REMOVING EXISTING DOCKER IMAGE: $image_name"
+        docker rmi -f "$image_name"
+    fi
 
     # check if .tar file exists
     if [ -f "$folder_name/$image_name.tar" ]; then
         echo "REMOVING EXISTING DOCKER IMAGE FILE."
         rm $folder_name/$image_name.tar
     fi
+
+    echo "HELP: "$folder_name/$image_name.tar" AND "$folder_name/Dockerfile""
 
     # check if Dockerfile exists
     if [ -f "$folder_name/Dockerfile" ]; then
